Create motion-test.py
This is a simplified version of the motion-to-turn-on-screen script.  This one seems a little more responsive (and less sophisticated). The motion.py script occasionally allows for a black/blank screen even while there is motion present.
It would be more ideal for the script to turn the screen on for ~30 seconds, and each time it detects a new motion it resets the timer. During that 30 second window we can reduce the sampling rate to every x seconds (maybe skip 15 seconds and then start sampling again), although I am not really sure this will save anything (the idea having been to conserve resources).
Additionally, the script should reset the timeout ('xset s 30'), and should automatically run at startup.